Fix +runDMFT: numerical precision problems in the stopping condition
Sometimes (quite rarely admittedly, but that's not the point) it would happen that the last point (Ustop) was disarded or, now that we had changed from a "while [] < []" condition to a "while [] ~= []" one, the loop could even last forever if the stopping condition failed due to numerical precision problems. So, most generally, we must avoid relying in real number equality to catch a stop (nothing along the lines of "stop if U==Ustop".

The simplest solution appears to be

	while abs(U-Ustop) > abs(Ustep)/2

with an obvious "min(abs(Ustep)/2)" modification for when Ustep is an array.