Fix stopping condition for runDMFT lines
Commit ff1c3bd7664a2824422c4cc7910d8554aeb73122 succesfully removed numerical precision problems on the stopping condition for all lines. Before it would happen occasionally that the last point is skipped (or, at an intermediate state, that the line would continue indefinitely if the condition is missed due to precision).

After that change the behavior has been stabilized in _always_ missing the last point :sweat_smile:

Here we fix the dumb error in the new, numerically stable, condition.

------------

En passant we also add a check on the relative sign of Ustop-Ustart and Ustep, so to avoid unintended infinite loops if they are different.

> A warning is raised to allert the user and Ustep is inverted in sign.