FIX: postDMFT.get_list on GNU Octave
I was wrong about substituting strings(N,1) with zeros(N,1)... it passes the linting but not the run time, for each element of a double-valued array has dimension (1) while the string I try to pass to it later has dimension (>1), being a char vector.

The solution is to work with cells, which apparently are perfectly handled by Octave for our purposes, i.e. list{1} and list(1) are different things, but give true if compared with strcmp :smile:

------

This commit is relevant to issue #2