Deal with missing readcell() in GNU Octave
The builtin function readcell() has been introduced in MATLAB very recently (since R2019a). Hence overcoming the possibility of it not parsing not only fixes GNU Octave compatibility [ #2 ] but also enhances generic code stability within MATLAB.

We donot really provide an alterantive implementation (meaning that we cannot mantain the very same functionality), but in case we cannot read the *info.ed files, we rise a warning for the user and just save the files for the lines in a generic format (observable#n and energy#n, with n=1,2,3...), so that the user can infere which is which by looking himself at the corresponding info file.