Replace split() with strsplit()
It is better supported in old MATLAB versions and in GNU Octave. Despite
being sizeably slower, the usage within the codebase is not performance
critical, so a straight substitution should be the right thing to do.

This should close for now issue #2 (but proper testing of the new plot
capabilities is highly warranted).