Fix CRAMERI colormaps
Ok, now I understand why the bug appeared subtle: there is no real diff
between the 1.7 and 1.8 version of Chad Green's CRAMERI function.
The 'fix' was just to change variable names inside the .mat archive, so
to have e.g. 'romeo' instead of 'romeO'.
This operation apparently has not been done with the updated ...7.0.mat
file so the bug has returned.

Here we provide a more robust fix, by moving the lower-case enforcing to
the try-catch construct. It's still not ideal, if we really want a case
insensitive behavior (which might just be a bad idea, given the original
naming convention by Crameri...), for it can handle whatever crazy case
choice for 'regular' maps, like 'BeRLin', but just the correct intended
case for the last-letter-is-capital maps, like 'grayC'. For now we stick
with this setup.

> Tested for all direct and reversed maps, as provided in the show-reel
  picture. Everything appears to work.

————————————————————————————————————————————————————————————————————————

A more general comment: we just broke the implicit do-not-touch-private
'rule'... so I wonder if we may dive in deeper and customize better the
original implementations by Green and Cobeldick? Coming soon I guess...