Update +postDMFT: let's put a NaN if cannot find a file
Too often it happens that a dmft_kinetic_energy.dat file is missing, for
that kinetic energy is computed only after the last loop, so it's easy
to skip in unfinished points. So here we provide a easy way to deal with
the issue: if the file is not present, let's just put a NaN in the array
which would result in a 'hole' in whatever plot function it goes into.

- easy to be detected either visually or with isnan
- does not break the whole line, which is *annoying*

We have added the very same functionality to every other *_line function
except observables, order parameters and energies, i.e. every line that
outputs a cell-array, for that we don't now a priori the cell-dimension,
and putting a single NaN won't prevent issues downstream when trying to
save files or plot stuff. Anyway all these quantities are computed early
at each loop, hence the likelihood of them be missing is quite low. :)