CRITICAL BUGFIX: runDMFT.refresh_point()
A longstanding problem that quite phantasmagorically has never arised,
until now, at the eve of my progress report, of course.

Up to now we used to move used.input* into input* before refreshing a
single-point calculation, so to ensure running with the same input state
(except new intended option provided with varargins). This is fine.

BUT...what happens if the process dies before finishing, what happens if
the DMFT code is KILLED BEFORE WRITING USED.INPUT AGAIN? It happens that
we corrupt ineluctably our directories. DAMN. And no, it's not unlikely.
It would happen anytime a refresh_line exceeds the runtime limit on hpc.

SO: let's just make a copy of used.input*, named input*, so to be super-
sure that whatever happens a VALID inputfile is stored and available for
subsequent runs (or even just as a record, damn).

————————————————————————————————————————————————————————————————————————
NB) For now I will also insert a spurious try-catch construct, aiming at
    recovering my now broken folder tree on Ulysses. I will remove these
    lines in the next, hopefully stable, update.