INTERMEDIATE: update runDMFt.refresh_line() [BREAKING API]
More flexible interface to define which points to refresh. We drop the
old ignConv flag and introduce a <mode> string.

• mode='all'     -> postDMFT.get_list is used to enter all the U-folders

• mode='conv'    -> U_conv.txt is used

• mode='nonconv' -> {get_list('U') - U_conv.txt} is used

This is expected to highly simplify the most common workflow: iterating
only nonconverged points to make them converge. For this reason we set
'nonconv' as the default value. Also if you want to iterate converged U-
values, most probably you want to pass varargins, which means you would
need to pass <mode> anyway. Please notice that this means the new API is
breaking back-compatibility, in that the default behavior has changed.