Add `postDMFT.compute_zeta`
It computes the quasi-particle weight on the REAL axis, as

           Z = 1 / ( 1 - d/dω[ReΣ(ω)]@ω=0 )

for a full Uloc line. There is no need to compute it on the Matsubara
axis since it is already done in the QcmPlab libraries, and the value
can be retrieved with either zeta_line() or observables_line().

SYNTAX: [Z_list,U_list] = postDMFT.compute_zeta(filename,window,U_LIST)

Z_list: returning all the quasiparticle weight values (forall U)
filename: an optional charvec, specifying which self-energy term to load
window: an optional window for the fit, abs(w)≤window [default 0.01]
U_LIST: an optional array of Hubbard interaction values: where to search