Add `postDMFT.get_pure_states`
[states, probs] = get_pure_states(suffix,check)

   states: a cell of pure-state vectors [states{i} is a complex array]
   probs: an array of pure-state probabilities [probs(i) = 𝓟(states{i})]
   suffix: an optional charvec / string, handling inequivalent filenames
   check: an optional boolean to activate the RDM crosscheck [default:F]

We may change the return type for 'states' (a plain matrix could be more
efficient, it depends on what we'll choose as our data structure for the
full pure_state_line / to optimize plotting of these states). We'll see.

The final goal is to plot something like: probs(i)*|states{i}(:)|^2, i.e.
the weight in which each Fock-space basis state enters the groundstate of
the impurity / cluster. This should clarify which configurations dominate
the physics of the system.