New version of `plotDMFT.pure_states` featuring a TREEMAP instead of PIE
This allows to display in the same graph both the information about p(i)
and |c{i}(j)|^2, i.e. what are the eigenvalues of the RDM defining the
probabilities associated to pure-states and how the basis states enter
these pure states. This is much more informative than the previous code
based on a pie chart. For a broader discussion on the limitations of pie
charts see also the commit message in 7b03f2aec3d12dde780c64a0ee20c40f25208a4c.