Give up legend in favor of a colorbar [`plotDMFT.spectral_stack`]
A legend would be way too much big: it won't fit in the plot, that's it.
A colorbar will convey less information and actually become misleading,
if the list of U values is not evenly spaced. But arguably the logic of
the plot itself is severely hindered if you pass an uneven ulist, so for
now at least we'd stay here.