Start allowing runners to handle generic "main" variables
Not only "U" lines, but whatever, passed, variable, lines.

Having it not breaking the old behavior is though. For dry_line we
managed to get the old behavior. For single_point we slightly break
back-compatibility, in that now passing the name of the variable is
mandatory (though, if you pass "U" you would recover the U=x.xxxxxx
directories, but the correct Uloc=x.xxxxxx fortran calls).

Somewhat tested with appropriate virtual drivers ("echo"): it seems
to work well.

TODO:

- Apply the patch to other line runners
- Correspondingly generalize the postDMFT scripts