Refactor `autost{o,e}p_line()` to take arbitrary main variable names
The autostop API follows now the vec_line convention, of passing generic
arrays of values, instead of the limited start:step:stop scheme. This is
in fact a breaking change, but the legacy use of this procedure has been
so limited that I do not deem this to be a problem.